Added a bit of error checking to setup.sh

+++ setup.sh
@@ -3,7 +3,6 @@
 # Check for current dotfiles in $HOME
 
 FILES="zshrc vimrc tmux.conf"
-#files="test1 test2 test3"
 SUFFIX=".bkup"
 BASEDIR=$(pwd)
 
@@ -11,9 +10,16 @@ for FILE in $FILES
 do	
 	if [ -f $HOME/.$FILE ]
 	then
-		echo "Found $HOME/.$FILE -> Moving to $HOME/.$FILE$SUFFIX"
+		# Check if a .bkup file already exists and exit if it does
+		if [ -f $HOME/.$FILE.bkup ]
+		then
+			echo "Found an existing .bkup file -- exiting"
+			exit 1
+		fi
+
+		# Rename existing files to .bkup
 		mv $HOME/.$FILE $HOME/.$FILE$SUFFIX
 	fi
-	echo "Creating symbolic link from $HOME/.$FILE to $BASEDIR/$FILE"
+	# Create symlink for each file in $FILES
 	ln -s $BASEDIR/$FILE $HOME/.$FILE
 done
